requirements.txt
accounts==0.2.4
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.0
amqp==2.6.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.1.0
bcrypt==4.0.1
billiard==3.6.4.0
bitarray==2.8.2
cachetools==5.3.1
cart==1.2.2
category==0.0.3
celery==4.4.7
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.6
cryptography==41.0.4
cssselect==1.2.0
cssutils==2.7.1
cytoolz==0.12.2
distlib==0.3.7
distro==1.8.0
Django==4.2.7
django-cors-headers==4.2.0
django-crispy-forms==2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
easyocr==1.7.1
ecdsa==0.18.0
email-validator==1.3.1
emails==0.5.15
eth-abi==4.2.1
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.0
eth-utils==2.2.1
exceptiongroup==1.1.3
fastapi==0.70.1
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
greenlet==3.0.0
gunicorn==20.1.0
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.1
httpx==0.25.1
idna==3.4
imageio==2.32.0
inflect==5.6.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==4.6.11
lazy_loader==0.3
lru-dict==1.2.0
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.1
openai==0.28.0
opencv-python-headless==4.8.1.78
packaging==23.2
parsimonious==0.9.0
passlib==1.7.4
Pillow==10.0.1
platformdirs==3.11.0
pluggy==0.13.1
premailer==3.10.0
protobuf==4.24.4
psycopg2-binary==2.9.9
py==1.11.0
pyasn1==0.5.0
pyclipper==1.3.0.post5
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pytest==5.4.3
python-bidi==0.4.2
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2023.3.post1
pyunormalize==15.0.0
pywin32==306
PyYAML==6.0.1
raven==6.10.0
razorpay==1.4.1
referencing==0.30.2
regex==2023.10.3
reportlab==4.0.4
requests==2.31.0
rlp==3.0.0
rpds-py==0.10.4
rsa==4.9
scikit-image==0.22.0
scipy==1.11.3
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
sqlparse==0.4.4
starlette==0.16.0
sympy==1.12
tenacity==6.3.1
tifffile==2023.9.26
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.11.8
vine==1.3.0
virtualenv==20.24.5
wcwidth==0.2.8
web3==6.10.0
websockets==11.0.3
yarl==1.9.2
